feat: migrate from Pages Api to App Router
