--- src/App/App.tsx
@@ -1,21 +1,0 @@
-// import { BrowserRouter, Route, Routes } from 'react-router-dom';
-// import './App.css';
-// import Results from '../Results';
-// import NotFound from '../NotFound';
-// import { ThemeSwitcher } from '../components/ThemeSwitcher/ThemeSwitcher';
-
-// function App() {
-//   return (
-//     <div>
-//       <ThemeSwitcher />
-//       <BrowserRouter>
-//         <Routes>
-//           <Route path="/" element={<Results />} />
-//           <Route path="*" element={<NotFound />} />
-//         </Routes>
-//       </BrowserRouter>
-//     </div>
-//   );
-// }
-
-// export default App;
