+++ src/ThemeContext.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, {
   createContext,
   useState,
